+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## NEXT
+- Added `isValid()` to `Schema` instance
 - Allow passing context to field properties functions (ex: `field.isRequired(context)`)
 
 ## v3.0.10